+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- Added configuration surface area to the OTel logs payload generator, in a
+  manner similar to OTel metrics.
 
 ## [0.27.0]
 ## Added